Use the explicit conversion operator.

+++ 3rdparty/stout/include/stout/net.hpp
@@ -159,7 +159,7 @@ inline Try<int> download(const std::string& url, const std::string& path)
   curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, nullptr);
   curl_easy_setopt(curl, CURLOPT_FOLLOWLOCATION, true);
 
-  FILE* file = fdopen(fd.get(), "w");
+  FILE* file = fdopen(static_cast<int>(fd.get()), "w");
   if (file == nullptr) {
     curl_easy_cleanup(curl);
     os::close(fd.get());
